size_t/ptrdiff_t transition part 5: the libast zakkaya
This is the fifth of the thickfold patch series, which enables ksh93
to operate within a 64-bit address space.

The parts of ksh93 affected by this commit are:
- The libast compatibility functions.
- The libast miscellaneous functions (e.g. optget and friends).
- The libast portability functions.

Remarks:
- The second argument of getfsstat can either be a long (FreeBSD)
  or a size_t (OpenBSD). This patch opts to cast the second argument
  as a size_t, which will produce a warning on FreeBSD and no warning
  on OpenBSD.
- _ast_assertfail and debug_fatal() end with abort(), so they were
  marked with the noreturn attribute.
- error_break() triggered a warning because it's given extern
  despite being uses as though it were static, so it was
  marked static to reflect actual usage.
  - typefix() was not extern, but was missing its static designation,
    so that has also been rectified.
- Removed cmdopen_20110505() to avoid bitrot.

Change in the number of warnings on Linux when compiling with clang using
-Wsign-compare -Wshorten-64-to-32 -Wsign-conversion -Wimplicit-int-conversion:
2,617 => 2,131 => 248 (progression from part 4 => part 5 => part 13)

Progresses https://github.com/ksh93/ksh/issues/592

+++ src/lib/libast/misc/recstr.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2023 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2025 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -70,7 +70,7 @@ recstr(const char* s, char** e)
 	case '+':
 	case '0': case '1': case '2': case '3': case '4':
 	case '5': case '6': case '7': case '8': case '9':
-		n = strton(s, &t, NULL, 0);
+		n = (int)strton(s, &t, NULL, 0);
 		if (n > 0 && t > (char*)s)
 		{
 			if (e)
@@ -107,7 +107,7 @@ recstr(const char* s, char** e)
 	case 'u':
 	case 'U':
 		while (*++s == ' ' || *s == '\t' || *s == ',');
-		n = strtol(s, &t, 0);
+		n = (int)strtol(s, &t, 0);
 		if (n < 0 || n > 15 || *t++ != '.')
 			break;
 		v = strtol(t, &t, 0);
@@ -176,7 +176,7 @@ recstr(const char* s, char** e)
 			case '0': case '1': case '2': case '3': case '4':
 			case '5': case '6': case '7': case '8': case '9':
 				v = 0;
-				a[n++] = strtol(s, &t, 0);
+				a[n++] = (int)strtol(s, &t, 0);
 				s = (const char*)t - 1;
 				continue;
 			}
